Undo Stucki's update to explicitly define KeyedByTypeCollection<T>
