+++ Microsoft.Azure.Amqp.Dnx/project.lock.json
@@ -12779,6 +12779,7 @@
     ".NETFramework,Version=v4.5.1": [
       "fx/System ",
       "fx/System.Runtime.Serialization ",
+      "fx/System.ServiceModel ",
       "fx/System.Transactions ",
       "fx/System.Xml "
     ],
