+++ Microsoft.Azure.Amqp.Uwp/Microsoft.Azure.Amqp.Uwp.csproj
@@ -271,7 +271,6 @@
     <Compile Include="..\Microsoft.Azure.Amqp\InternalBufferManager.cs" />
     <Compile Include="..\Microsoft.Azure.Amqp\Interop\UnsafeNativeMethods.cs" />
     <Compile Include="..\Microsoft.Azure.Amqp\IteratorAsyncResult.cs" />
-    <Compile Include="..\Microsoft.Azure.Amqp\KeyedByTypeCollection.cs" />
     <Compile Include="..\Microsoft.Azure.Amqp\PartialTrustHelpers.cs" />
     <Compile Include="..\Microsoft.Azure.Amqp\ResourcesGeneric.Designer.cs">
       <DependentUpon>ResourcesGeneric.resx</DependentUpon>
